third_party/WebKit/LayoutTests/printing/printing-transition-ignore-media-print.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
#target {
  height: 250px;
  transition: all 10s;
}

@media print {
  #target {
    height: 800px;
  }
}
</style>

<div id="target"></div>

<script>
test(function() {
  if (window.internals) {
    assert_equals(getComputedStyle(target).height, "250px");
    assert_equals(internals.numberOfPages(500, 500), 2);
    assert_equals(getComputedStyle(target).height, "250px");
  }
}, "This test checks that transition are ignored when in @media print");
</script>
