third_party/blink/web_tests/flag-specific/highdpi/virtual/layout_ng_printing/printing/printing-transition-ignore-media-print-expected.txt

This is a testharness.js-based test.
FAIL This test checks that transition are ignored when in @media print assert_equals: expected 2 but got 1
Harness: the test ran to completion.
